+++ CHANGELOG.md
@@ -12,18 +12,16 @@ Full documentation for MIGraphX is available at
 
 ### Changed
 
+* Refactored instruction output alias to return a vector of aliases (#4540).
 
 ### Resolved issues
 
 * Fixed a bug with operators `pack_fp4`, `unpack_fp4`, and the `fuse_mlir` pass handling non-standard input shapes (#4560).
 * Fixed an issue in `propagate_precision` pass where precision for div operations could be incorrectly propagated across type boundaries (e.g., from integral to floating-point) (#4561).
 
-
 ### Optimized
 
-
 ### Removed
-
 
 ## MIGraphX 2.15 for ROCm 7.2.0